refactor: change the name and path of drawerhooks.js
